+++ CHANGELOG.md
@@ -9,6 +9,8 @@ CHANGELOG
 - Upgrade jmespath to ~=1.0 (from ~=0.10).
 - Upgrade tabulate to <=0.9.0 (from <=0.8.10).
 - Add support for p6-b300 instances for all OSs except AL2.
+- Add permission `cloudwatch:PutMetricData` to the head node policy so that clustermgtd is able to emit metrics.
+- Add alarm on missing clustermgtd heartbeat.
 
 **BUG FIXES**
 - Add validation to block updates that change tag order. Blocking such change prevents update failures.